Integrate the following post-bzr (post 2.18.3) patch from the mainline distcc sources (http://bazaar.launchpad.net/%7Embp/distcc/trunk/):
   % bzr log -r3
   message: merge avahi patch from Lennart Poettering

(This patch comes from http://0pointer.de/public/distcc-avahi.patch,
which is the version of the patch I applied, since the one in the bzr
repository was missing the new files from the avahi patch.)

I've also manually updated ChangeLog, which is not automatically
updated by bzr the same way it is by arch.

I've also added a new m4 directory, including non-standard autoconf
macros that we use.  For now, it only includes ACX_PTHREAD (added by
the avahi patch) which I got from 0pointer.de.  See the notes at the
top of m4/acx_pthread.m4 for more details.

I removed the change to Makefile.in that added in distccmon-gnome.1.
This file is not actually added in as part of this patch (it seems to
be part of a separate patch, at
http://patches.ubuntu.com/d/distcc/extracted/02_distccmongnome_man.dpatch),
and may have gotten in here by accident.

Tested by running 'make maintainer-check'.  However, I did not test
with avahi installed, so I don't know how well that works.

Reviewed by fergus@google.com and klarlund@google.com



git-svn-id: http://distcc.googlecode.com/svn/trunk@16 01de4be4-8c4a-0410-9132-4925637da917