Annoyingly, different versions of gcc are inconsistent in how they treat a non-existent #include file when invoked with "-MMD": some versions treat it as an error (rc 1), some as a warning (rc 0).  When distcc is responsible for preprocessing (_server_options includes 'cpp'), we need to figure out which our gcc does, in order to verify distcc is doing the same thing.
Before this change, we just hard-coded in one sort of gcc behavior
(that it always warns).  This change actually tries gcc out to see
which it does, to make sure distcc does the same.

Tested by running 'make pump-maintainer-check' on both Linux gcc 4.0.3
(which warns) and FreeBSD gcc 3.4.4 (which errors).

Reviewed by fergus@google.com


git-svn-id: http://distcc.googlecode.com/svn/trunk@186 01de4be4-8c4a-0410-9132-4925637da917