Get rid of 'readlink -f', which is a Gnuism.
The purpose of the readlink -f was to give the absolute path of the
installation; that happens to reveal the installation date given the way we set
up distcc-pump at Google.  But this does not work under FreeBSD.  The
installation information can be communicated in different ways (for example by
explicit actions of the build system).

Reviewers: fergus@google.com, csilvers@google.com


git-svn-id: http://distcc.googlecode.com/svn/trunk@189 01de4be4-8c4a-0410-9132-4925637da917