On some systems (FreeBSD), it looks like connect() can return EINVAL. This isn't documented, but there's some talk about it on the web, eg    http://mail-index.netbsd.org/netbsd-bugs/2003/08/18/0002.html
It looks like the cause is that the first connect fails for whatever
reason, and then subsequent ones fail with EINVAL.

I've detected that in the NoDetachDaemon_Case, both on FreeBSD and OS
X machines.  The symptom is that sock.connect_ex(('127.0.0.1',
self.server_port) continually returns EINVAL, and no amount of time
will fix things.

I'm not exactly sure why, but I figure one possible problem is the
connect is happening before distccd has started listening, and that
causes problems.  So I put in a bit of a delay to give distccd time to
start up.  I figure it can't hurt, and may well help.

Tested by running 'make check' on a ubuntu machine (not subject to
this problem) and a freebsd machine (which sometimes is).  It
certainly didn't hurt, though I can't tell if it solved the problem
for good.

Reviewed by fergus@google.com


git-svn-id: http://distcc.googlecode.com/svn/trunk@222 01de4be4-8c4a-0410-9132-4925637da917