1) Fix up a PATH= assignment that would fail if builddir had spaces in it (because the output of `cd "$builddir" && pwd` was not being quoted).  As a bonus, make the PATH= fit on one line, which makes some shells happier.
2) Fixed up some install targets.  On some systems, tar without -f
reads from stdin, but that's non-standard (it should read from some
tape-drive device in /dev in that case).  Add -f- to $(TAR) in that
case, which should work everywhere.

3) cp -a isn't supported on all systems.  I checked what automake
does, and it emits cp -pR, so that's what I do too.  We don't have any
internal symlinks that I need to worry about -P or anything like
that.

Tested by running 'make installcheck' on a freebsd 6.0 system.

Reviewed by fergus@google.com


git-svn-id: http://distcc.googlecode.com/svn/trunk@239 01de4be4-8c4a-0410-9132-4925637da917