Add a comment.  From klarlund@google.com:
"Assume that:

(1)  'gcc' is a link to 'distcc' earlier on PATH than the real gcc.
(2)  gcc is invoked; in fact that invokes distcc, which will rewrite
the PATH in src/climasq.c before invoking 'gcc' again.

That will make calls by distcc such as 'gcc -E' for preprocessing work
OK, but the include processor will in fact be calling the 'distcc' as
'gcc' again! But because the input argument is an I/O redirection,
distcc will then correctly call 'gcc'. In args.c it fails to recognize
an input file."


git-svn-id: http://distcc.googlecode.com/svn/trunk@245 01de4be4-8c4a-0410-9132-4925637da917