Modifications to the pump and test framework for cygwin.
1) find_c_extension looks for .dll files in addition to .so files.

2) Likewise, testdistcc.py has to check whether binary names might end
in .exe.  Usually that doesn't matter, but sometimes (eg when running
cp on them), it does.

3) testdistcc.py adds a routine to detect PE binaries.  These, like
dwarf binaries, do not always give reproducible output on the same
input, so take that into account in the gdb tests.

4) In the lsdistcc test, we assumed that when the ping failed,
127.0.0.2 isn't available.  But it's also possible ping is just broken
(or different) on that machine.  So in the failed-ping case, we no
longer make any assumption about 127.0.0.2, one way or the other.

Tested by running 'make check' on cygwin.  Most every test passes!

Reviewed by fergus@google.com


git-svn-id: http://distcc.googlecode.com/svn/trunk@257 01de4be4-8c4a-0410-9132-4925637da917