There's cygwin-specific code in exec.c, but it seems to assume that when running under cygwin, we'll be using a native windows app to compile.  But sometimes we're using a cygwin app (eg /bin/gcc). Trying to use native-windows forking routines with cygwin apps causes a segfault (and takes a while to do it); we need to use unix forking routines there.
With this change, when we're asked to spawn a sub-process, we try to
guess whether the sup-process app is windows or cygwin, based on its
path location (this benefits from using absolute paths):
   /bin/gcc
vs
   C:\Program Files\Microsoft Visual Studio .NET 2003\Vc7\bin\cl.exe
or even
   C:/Program Files/Microsoft Visual Studio .NET 2003/Vc7/bin/cl.exe

If we decide it's a windows app but are wrong, then the
windows-forking code should eventually fail with "file not found"; in
that case, we fall through to the normal unix-fork case, to give
cygwin a chance to find the binary.

I also took the opportunity to clean up some of the code (handles
should be initialized to ILLEGAL_HANDLE values, not NULL).  I also
replaced the crazy permissions the code was asking for before, with
normal GENERIC_READ/GENERIC_WRITE.  This allows the code to work when
the tmpdir is c:\windows, which was failing before.

Tested on cygwin using /bin/gcc as the compiler, like so:
   ./distcc /bin/gcc -c testtmp.c -o testtmp.o
where testtmp.c is a simple "hello world" program.

I also tested by running 'make check'.  With this change, 'make check'
passes on cygwin!  (Using gcc to do compilations.)

Reviewed by fergus@google.com


git-svn-id: http://distcc.googlecode.com/svn/trunk@282 01de4be4-8c4a-0410-9132-4925637da917