Set-up masquerading for benchmarks.
Follow the advice of the distcc man page for executing the benchmarks to not
rely on CC and CXX variables by also routing naked calls of 'cc', 'gcc', etc to
a 'farm' of tiny shell scripts.

Also, provide a little wrapper script that does this for people who have to
debug a benchmark build.

This change is convenient for running the benchmark on the linux kernel,
because building the kernel uses naked 'gcc' calls.

Testing:

  Building the linux kernel.


git-svn-id: http://distcc.googlecode.com/svn/trunk@373 01de4be4-8c4a-0410-9132-4925637da917