Make the Linux 2.6 kernel build with distcc-pump mode.
Add a new piece of command line parse information: send_systemdirs.  This
boolean is normally false.  It is true if there is a -isystem option with a
default system directory (one known to the compiler).

With send_systemdirs true, the compiler headers of system headers are sent to
the servers and mounted under the server root.  The isystem option is as usual
rewritten to be relative to the root.  Without this flag setting, distcc quickly
decides that pump mode is not viable because remote compilations fail.

Also, fixed the comment:
  # TODO(klarlund): Make mechanism for handling -U, -undef options, along with
  # default symbols.

(-D options are handled.)

---------------

TESTS: make pump-maintainer-check
       make include-server-maintainer-check
       make 'linux kernel' in a directory made by the benchmark script (as shown below)

      DISTCC_EMAILLOG_WHOM_TO_BLAME=klarlund@google.com DISTCC_ENABLE_DISCREPANCY_EMAIL=1 INCLUDE_SERVER_ARGS='-d1 -t --email --unsafe_absolute_includes --send_systemdirs --stat_reset_triggers=include/linux/*' DISTCC_HOSTS="--randomize `/home/klarlund/svn-distcc/distcc/lsdistcc -P3 -k100 -d`" PATH=/home/klarlund/svn-distcc/distcc:$PATH  ./masquerade pump make -j120

REVIVIEWER: csilvers@google.com

TRIVIA: the about 1000 files in the kernel build involve on average 400
headers. The build time on a quad-cpu machine drops from about 1m40 s to 1m20s.
With pump-mode, the machine is underutilized judging from the total sys + user
time, which is around only 2m20s.

NOTE: this is a successor to 'g6n8', which simply introduced a command line
option to force sending of system files.  A bug in gvn did not allow me to
resuse that change list.


git-svn-id: http://distcc.googlecode.com/svn/trunk@393 01de4be4-8c4a-0410-9132-4925637da917