Add critique of pump-mode when include server shuts down.
Print a message like:

__________Warning: 1 pump-mode compilation(s) failed on server, but succeeded locally.

when compile discrepancies have occurred.  Currently, a user may think that all
is well because we do not terminate a build when discrepancies occur and the
warning messages occurring in the middle of a build may easily be overlooked.

REVIEW: csilvers@google.com
TESTS:  make linux-kernel without stat reset triggers to observe message



git-svn-id: http://distcc.googlecode.com/svn/trunk@404 01de4be4-8c4a-0410-9132-4925637da917