Tidy-up testing framework to overcome recently introduced bug.
Testing on cygwine revealed that as tests were being issued the temporary
testing directory would suffer from longer and longer prefixes until a 'filename
too long' error would stop the party.  This is result of an earlier attempt to
avoid too much printing to stdout during tests, spefically during recursive test
invocations.

The problem is that the teardown mechanism of the comfychair enviroment is not
expressed -- by any name at all (it could be 'teardown' for example). So, I
added a name --- but because there's no expectation that the TestCase class
defined there has its teardown method called from derived classes --- I chose
the name CleanUps, which corresponds to the private variable.

I also made an abstraction for running a test. This new Comfychair.runtest
function can the conveniently by called to carry out a subtest.

With this CL, 'make pump-maintainer-check' now works on cygwin.


git-svn-id: http://distcc.googlecode.com/svn/trunk@434 01de4be4-8c4a-0410-9132-4925637da917