The missing include_server(1) man page and additional material about the basic assumptions of distcc-pump added to man/distcc.1 as well. The BUGS section of distcc.1 has been updated.
A spelling error correction is made to man/distcc.1. 

The pump.in script is changed so that it does not introduce other certain
defaults than those mentioned in the manual (which are those of the of include
server).

Finally, I changed the order of the options in the include server source code to
be alphabetical in accordance with include_server(1) man page.

Reviewers: fergus, csilvers.


git-svn-id: http://distcc.googlecode.com/svn/trunk@481 01de4be4-8c4a-0410-9132-4925637da917