Fix a bug reported by Aaron P. Perez <Aaron.Perez@hospira.com> where "make install" was failing on Cygwin.
The problem was due to the install.log file generated by Python's
distutils being generated in DOS text mode format rather than
Unix text mode format.  The fix is to use 'sed' to convert
DOS text format to Unix text format when generating install.log.

Note: this fix is slightly different than the original patch
that I posted to the distcc mailing list.  This patch is better
because it fixes the install.log file, rather than just the
grep command; this is important because the install.log file
is also used for "make uninstall".

Reviewers: Craig Silverstein


git-svn-id: http://distcc.googlecode.com/svn/trunk@549 01de4be4-8c4a-0410-9132-4925637da917