Add a test of C++ compilation, and refactor the tests of Objective-C and Objective-C++ to share the common code.
Also, fix some bugs in the tests of Objective-C and Objective-C++.
These tests were never being activated, even when the right compiler is
installed, because in the test compilation (not using distcc) it was
compiling with "-o /dev/null" and then testing that testtmp.o exists,
which it never did.  There was also a missing "std::" in the Objective-C++
test.  I still haven't run the Objective-C++ test, because I still haven't
figured out how to install Objective-C++, but it's pretty similar to
the C++ and Objective-C tests, and both of those now run and pass.

Reviewers: Craig Silverstein


git-svn-id: http://distcc.googlecode.com/svn/trunk@587 01de4be4-8c4a-0410-9132-4925637da917